+++ lib/main/dial_widget.dart
@@ -3,7 +3,6 @@ import 'timer_painter.dart';
 import 'main_module.dart';
 import 'central_button.dart';
 import 'session_digit.dart';
-import '../app/models.dart';
 import 'minute_digit.dart';
 
 class DialWidget extends StatelessWidget {
@@ -16,14 +15,14 @@ class DialWidget extends StatelessWidget {
   final int elapsedTime;
   final VoidCallback onTapListener;
   final bool paused;
-  final int minute;
-  final TimeOfDay startTime;
+  final DateTime currentTime;
+  final DateTime startTime;
 
   DialWidget({
     this.elapsedTime,
     this.onTapListener,
     this.paused,
-    this.minute,
+    this.currentTime,
     this.startTime,
   }):assert(onTapListener != null);
 
@@ -43,7 +42,7 @@ class DialWidget extends StatelessWidget {
       new Transform(
           transform:
               new Matrix4.translationValues(size.width / 2, DIAL_CENTER, 0.0),
-              child: new MinuteDigit(minute: minute, radius: DIAL_RADIUS - 10.0),
+              child: new MinuteDigit(currentTime: currentTime, radius: DIAL_RADIUS - 10.0),
         ),
       new Transform(
         transform: new Matrix4.translationValues(
@@ -53,27 +52,10 @@ class DialWidget extends StatelessWidget {
       new Transform(
           transform:
               new Matrix4.translationValues(size.width / 2, DIAL_CENTER, 0.0),
-          child: new CustomPaint(
-            size: new Size(radius * 2, radius * 2),
-            painter: new SessionDigitPainter(
-              sections: <Section>[
-                new Section(
-                  length: 1000 * 60 * 25,
-                  foregroundPaint: workSectionCompletePaint,
-                  backgroundPaint: workSectionIncompletePaint,
-                  sessionType: SectionType.WORK,
-                ),
-                new Section(
-                  length: 1000 * 60 * 5,
-                  foregroundPaint: breakSectionCompletePaint,
-                  backgroundPaint: breakSectionIncompletePaint,
-                  sessionType: SectionType.BREAK,
-                )
-              ],
-              dialOuterRadius: 135.0,
-              dialInnerRadius: 50.0,
-              elapsedLength: elapsedTime,
-            ),
+          child: new SessionDigit(
+            radius: radius,
+            elapsedTime: elapsedTime,
+            startTime: startTime,
           ),
         ),
