+++ lib/main/main_widget.dart
@@ -10,8 +10,14 @@ class MainWidget extends StatefulWidget {
 class _MainWidgetState extends State<MainWidget> {
   var elapsedTime = 0;
   bool isPaused = true;
-  int minute = 0;
-  TimeOfDay startTime;
+  DateTime currentTime;
+  DateTime startTime;
+
+  @override
+  void initState() {
+    startTime = new DateTime.now();
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -21,20 +27,20 @@ class _MainWidgetState extends State<MainWidget> {
           if(!isPaused) {
             elapsedTime +=100;
           }
-          minute = new TimeOfDay.now().minute;
+          currentTime = new DateTime.now();
         }),
       );
     return new DialWidget(
       elapsedTime: elapsedTime,
       onTapListener: _handleOnTap,
       paused: isPaused,
-      minute: minute,
+      currentTime: currentTime,
       startTime: startTime,
     );
   }
 
   void _handleOnTap () => setState((){
     isPaused = !isPaused;
-    startTime = new TimeOfDay.now();
+    startTime = new DateTime.now();
   });
 }
