+++ lib/main/minute_digit.dart
@@ -4,29 +4,30 @@ import 'dart:math';
 
 class MinuteDigit extends StatelessWidget {
 
-  final int minute;
+  final DateTime currentTime;
   final double radius;
-  MinuteDigit({Key key,this.minute,this.radius}):super(key:key);
+  MinuteDigit({Key key,this.currentTime,this.radius}):assert(currentTime != null),super(key:key);
 
   Widget build(BuildContext contex) {
     return new CustomPaint(
       size: new Size(radius *2, radius *2),
-      painter: new _MinuteDigitPainter(minute)
+      painter: new _MinuteDigitPainter(currentTime)
     );
   }
 }
 
 class _MinuteDigitPainter extends CustomPainter {
-  final int minute;
-  _MinuteDigitPainter(this.minute);
+  final DateTime currentTime;
+  _MinuteDigitPainter(this.currentTime);
   bool shouldRepaint(_MinuteDigitPainter oldPainter) {
-    return minute != oldPainter.minute;
+    return currentTime != oldPainter.currentTime;
   }
 
   @override
   void paint(Canvas canvas, Size size) {
+    var timestamp = currentTime.minute * 60 + currentTime.second;
     var p1 = new Offset(0.0,0.0);
-    var p2 = new Offset(cos(DEFAULT_ANGLE_CORRECTION + minute / 30 * PI) * size.width/2,sin(DEFAULT_ANGLE_CORRECTION + minute / 30 * PI) * size.width/2);
+    var p2 = new Offset(cos(DEFAULT_ANGLE_CORRECTION + timestamp / 60 / 30 * PI) * size.width/2,sin(DEFAULT_ANGLE_CORRECTION + timestamp / 60 / 30 * PI) * size.width/2);
     canvas.drawLine(p1, p2, minuteDigitPaint);
   }
 }
